refectored animal quiz modal
